+++ HISTORY.md
@@ -2,6 +2,8 @@
 
 ## 0.1.6 (2020-??-??)
 
+* Changed behaviour: No reconnect of the serial connection will be performed for each request,
+  because each `login()` call during a request will automatically try a reconnect on failure.
 * Clean-up of `setup.py` and `MANIFEST.in`.
 
 ## 0.1.5 (2020-03-29)